+++ CHANGELOG.md
@@ -2,19 +2,25 @@
 
 All Notable changes to `newsletter-form` will be documented in this file
 
+## 2.0.0
+
+### Changed
+- Extra option for submit button
+- Disable button on submit, enable on response
+
 ## 1.0.3
 
-## Fixed
+### Fixed
 - Bug when using no custom options
 
 ## 1.0.2
 
-## Fixed
+### Fixed
 - Improved readability
 
 ## 1.0.1
 
-## Fixed
+### Fixed
 - Better syntax
 
 ## 1.0.0